requirements.txt
pysam>=0.19.0
numpy>=1.20.0
biopython>=1.79
click>=8.0.0